requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.14.3; python_full_version >= '3.7.0'
brotli==1.2.0
certifi==2026.1.4; python_version >= '3.7'
charset-normalizer==3.4.4; python_version >= '3.7'
click==8.3.1; python_version >= '3.10'
feedparser==6.0.12; python_version >= '3.6'
filelock==3.20.3; python_version >= '3.10'
greenlet==3.3.0; python_version >= '3.10'
idna==3.11; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
joblib==1.5.3; python_version >= '3.9'
lxml[html-clean]==6.0.2; python_version >= '3.8'
lxml-html-clean==0.4.3
markupsafe==3.0.3; python_version >= '3.9'
newspaper4k==0.9.4.1; python_version >= '3.10'
nltk==3.9.2; python_version >= '3.9'
pillow==12.1.0; python_version >= '3.10'
playwright==1.57.0; python_version >= '3.9'
pyee==13.0.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pyyaml==6.0.3; python_version >= '3.8'
regex==2026.1.15; python_version >= '3.9'
requests==2.32.5; python_version >= '3.9'
requests-file==3.0.1
sgmllib3k==1.0.0
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
soupsieve==2.8.1; python_version >= '3.9'
tldextract==5.3.1; python_version >= '3.10'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
urllib3==2.6.3; python_version >= '3.9'
